Moving substance into substance.rs
